Update Weeks 5 - 10
Over these weeks my focus was mostly on updating player movement and getting part of report worked on in terms of literature review. For the prototype I have managed to implement the enemies into the procedural generation algorithm. The way it works at the moment is when the rooms are created, an algorithm takes into account the number of rooms in total and calculates at least one enemy per each room in the dungeon (that does not mean the enemy will be an enemy per each room though). I also worked on changing the movement system for the player to get more smooth RPG kind of style movement rather than turn based one.