Fixed The Dungeon Generation bug
I updated the player animation code. Implemented a title screen with transition to game world. Fixed the dungeon generation code, the bug was in the "Direction" enum in Corridor.cs, where I tried to add extra directions in wrong order (I was thinking of using it along with Player Animation/Movement Code, so to have EastWest direction, NorthWest, etc.) I decided to create a separate enum for player's direction underneath it and it fixed that problem. Basically, the corridor tried to be generated for example "North East" side (Which would be impossible to create it in a "square" corridor, due to Enums being in wrong order than in the initial dungeon generation.)